+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Upgraded swagger-ui to v2.1.1 - [@dblock](https://github.com/dblock).
 * Grape-swagger 0.7.2 is no longer supported - [@dblock](https://github.com/dblock).
 * Implemented RuboCop, Ruby-style linter - [@dblock](https://github.com/dblock).
+* [#31](https://github.com/ruby-grape/grape-swagger-rails/pull/31): Support Swagger-UI docExpansion option - [@maruware](https://github.com/maruware).
 * Your contribution here.
 
 ### 0.1.0 (February 5, 2015)